segregating schema type component
